:recycle: refactor: object Constants 제거 및 object TimeConverter 추가
- object Constants 제거
  - 캡슐화 X
- object TimeConverter 정의하여 내부에 변수 및 변환 함수 추가

Issue #76
