+++ presentation/src/main/kotlin/com/woory/presentation/ui/creatingpromise/LocationSearchResult.kt
@@ -2,7 +2,6 @@ package com.woory.presentation.ui.creatingpromise
 
 import com.woory.presentation.model.GeoPoint
 
-
 data class LocationSearchResult(
     val name: String,
     val address: String,
